add date_time support to the route
